src/godri/e2e_tests/test_mcp_e2e.py
"""End-to-end tests for MCP server functionality."""

import asyncio
import json
import tempfile
import pytest
import logging

# Configure logging for tests
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class TestMcpE2E:
    """End-to-end tests for MCP server operations."""

    @pytest.fixture(autouse=True)
    async def setup(self):
        """Setup test environment."""
        # TODO: Initialize MCP server after refactoring
        self.mcp_server = None

    @pytest.mark.asyncio
    async def test_mcp_drive_tools(self):
        """Test MCP Drive tools end-to-end."""
        try:
            # TODO: Implement after refactoring MCP server
            # 1. Initialize MCP server
            # 2. Call drive_search tool
            # 3. Call drive_upload tool with test file
            # 4. Call drive_download tool
            # 5. Verify operations work

            logger.info("Test placeholder: mcp_drive_tools")
            assert True  # Placeholder

        except Exception as e:
            logger.error("MCP Drive tools test failed: %s", str(e))
            raise

    @pytest.mark.asyncio
    async def test_mcp_docs_tools(self):
        """Test MCP Docs tools end-to-end."""
        try:
            # TODO: Implement after refactoring MCP server
            # 1. Create document via MCP
            # 2. Update document content via MCP
            # 3. Read document via MCP
            # 4. Verify operations work

            logger.info("Test placeholder: mcp_docs_tools")
            assert True  # Placeholder

        except Exception as e:
            logger.error("MCP Docs tools test failed: %s", str(e))
            raise

    @pytest.mark.asyncio
    async def test_mcp_translate_tools(self):
        """Test MCP Translation tools end-to-end."""
        try:
            # TODO: Implement after refactoring MCP server
            # 1. Call translate_text tool
            # 2. Verify translation works

            logger.info("Test placeholder: mcp_translate_tools")
            assert True  # Placeholder

        except Exception as e:
            logger.error("MCP Translation tools test failed: %s", str(e))
            raise


if __name__ == "__main__":
    # Run tests directly
    test_instance = TestMcpE2E()
    asyncio.run(test_instance.test_mcp_drive_tools())
