pyproject.toml
[project]
name = "godri"
version = "0.2.0"
description = "Google Drive CLI tool"
authors = [{name = "Sebastien MORAND", email = "sebastien.morand@loreal.com"}]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "google-cloud-translate>=3.0.0",
    "colorlog>=6.0.0",
    "aiofiles>=24.0.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "mcp>=1.0.0",
    "uvicorn>=0.27.0",
    "starlette>=0.47.1",
    "fastapi>=0.116.1",
    "google-cloud-speech>=2.33.0",
    "mutagen>=1.47.0",
    "typer>=0.16.0",
]

[project.scripts]
godri = "godri.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/godri"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "bandit>=1.8.6",
]
